+++ server/handlers.go
@@ -7,6 +7,7 @@ import (
 	"net/http"
 	"strconv"
 
+	"github.com/gogogomoku/gomoku_v2/ai"
 	"github.com/gogogomoku/gomoku_v2/arcade"
 	"github.com/gogogomoku/gomoku_v2/arcade/match"
 	"github.com/gogogomoku/gomoku_v2/board"
@@ -110,6 +111,7 @@ func PostMoveHandler(w http.ResponseWriter, r *http.Request) {
 		http.Error(w, errMsg, http.StatusBadRequest)
 		return
 	}
+	match.Suggestion = ai.GetSuggestion(match.Board, match.History[len(match.History)-1], match.GetOpponent(player))
 
 	_ = json.NewEncoder(w).Encode(
 		truncateHistory(match),
@@ -134,6 +136,13 @@ func PostUnapplyMoveHandler(w http.ResponseWriter, r *http.Request) {
 		errMsg := fmt.Sprintf("Bad request: %s", err)
 		http.Error(w, errMsg, http.StatusBadRequest)
 		return
+	}
+	// Recalculate suggestion for player
+	if len(match.History) == 0 {
+		match.Suggestion = &board.Position{X: board.SIZE / 2, Y: board.SIZE / 2}
+	} else {
+		lastMove := match.History[len(match.History)-1]
+		match.Suggestion = ai.GetSuggestion(match.Board, lastMove, lastMove.Player)
 	}
 
 	_ = json.NewEncoder(w).Encode(
